Fix compilation with pre-compiled header

+++ src/dsp/Delay/VariableDelay.cpp
@@ -1,10 +1,10 @@
 #include "VariableDelay.h"
 
 VariableDelay::VariableDelay (size_t size) :
-    l0Delay (size),
-    l1Delay (size),
-    l3Delay (size),
-    l5Delay (size)
+    l0Delay (static_cast<int> (size)),
+    l1Delay (static_cast<int> (size)),
+    l3Delay (static_cast<int> (size)),
+    l5Delay (static_cast<int> (size))
 {}
 
 void VariableDelay::setDelay (float newDelayInSamples)
